fix(search): address PSO code review feedback
- Add errno check to sr_optimizer_read_real_env for overflow detection
- Add sr_pso_iter_limit = MAX_ITER_PSO to else branch for defaults
- Change sr_pso_cfg seed type from unsigned long long to uint64_t
- Add stdint.h include to search_func.h
- Add PSO algorithm description with LaTeX equations to csearch.rst
- Add default values and recommended ranges to PSO CLI options

+++ src/include/search_func.h
@@ -17,6 +17,7 @@ extern "C" {
 #define SEARCH_FUNC_H
 
 #include <stdio.h>
+#include <stdint.h>
 
 /**
  * @file search_func.h
@@ -118,7 +119,7 @@ typedef struct sr_pso_cfg {
   real c1;
   real c2;
   real v_max;
-  unsigned long long seed;
+  uint64_t seed;
 } sr_pso_cfg;
 
 /**
